# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: International payments collection, link, payout, settlement and operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment-links:
    get:
      operationId: listPaymentLinks
      tags: [payment-links]
      summary: List payment links
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Payment links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLinkList"
    post:
      operationId: createPaymentLink
      tags: [payment-links]
      summary: Create a fixed-price or customer-entered payment link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentLinkInput"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
  /payment-links/{id}:
    patch:
      operationId: updatePaymentLink
      tags: [payment-links]
      summary: Update or pause a payment link
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentLinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
    delete:
      operationId: deletePaymentLink
      tags: [payment-links]
      summary: Archive a payment link
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Archived
...
  /payouts:
    get:
      operationId: listPayouts
      tags: [payouts]
      summary: List payouts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, approved, completed, rejected, failed]
        - name: currency
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Payouts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutList"
    post:
      operationId: createPayout
      tags: [payouts]
      summary: Send a payout through an available Payzaapi payout method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutInput"
      responses:
        "201":
          description: Payout submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
  /payout-methods:
    get:
      operationId: listPayoutMethods
      tags: [payouts]
      summary: Discover methods, minimums and fees from Payzaapi
      parameters:
        - name: currency
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Current payout methods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutMethodsResponse"
  /banks:
    get:
      operationId: listBanks
      tags: [payouts]
      summary: List banks for a country
      parameters:
        - name: country
          in: query
          schema: { type: string, default: KE }
      responses:
        "200":
          description: Banks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankList"
  /settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      summary: List T+3 expected settlements
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, due, settled, held]
        - name: currency
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Settlement ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementList"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers with payment totals
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
  /webhook-events:
    get:
      operationId: listWebhookEvents
      tags: [webhooks]
      summary: Review incoming provider webhooks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [processed, failed, ignored]
      responses:
        "200":
          description: Webhook events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEventList"
  /webhook-events/{id}/replay:
    post:
      operationId: replayWebhookEvent
      tags: [webhooks]
      summary: Reprocess a stored webhook event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Replayed event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEvent"
...
  /v1/payment-links:
    get:
      operationId: listDeveloperPaymentLinks
      tags: [developer-api]
      summary: List payment links belonging to the API key's merchant
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Merchant payment links
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentLinkList" }
    post:
      operationId: createDeveloperPaymentLink
      tags: [developer-api]
      summary: Create a payment link for the API key's merchant
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentLinkInput" }
      responses:
        "201":
          description: Payment link
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentLink" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: merchant API key
  schemas:
...
    PaymentLink:
      type: object
      required: [id, slug, name, amountType, currency, status, url, paidCount, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        amountType: { type: string, enum: [fixed, customer_choice] }
        amount: { type: ["number", "null"] }
        currency: { type: string }
        status: { type: string, enum: [active, paused, archived] }
        url: { type: string, format: uri }
        paidCount: { type: integer }
        totalPaid: { type: number }
        expiresAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    PaymentLinkList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PaymentLink" }
    PaymentLinkInput:
      type: object
      required: [name, amountType, currency]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        amountType: { type: string, enum: [fixed, customer_choice] }
        amount: { type: number, exclusiveMinimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        expiresAt: { type: string, format: date-time }
    PaymentLinkUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string, enum: [active, paused] }
        expiresAt: { type: ["string", "null"], format: date-time }
...
    Payout:
      type: object
      required: [id, reference, amount, currency, method, accountName, status, provider, createdAt]
      properties:
        id: { type: integer }
        merchantId: { type: ["integer", "null"] }
        reference: { type: string }
        amount: { type: number }
        fee: { type: ["number", "null"] }
        netAmount: { type: ["number", "null"] }
        currency: { type: string }
        method: { type: string }
        accountName: { type: string }
        maskedAccount: { type: string }
        status:
          type: string
          enum: [pending, processing, approved, completed, rejected, failed]
        provider: { type: string, enum: [payzaapi, paystack] }
        createdAt: { type: string, format: date-time }
    PayoutList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Payout" }
    PayoutInput:
      type: object
      required: [amount, currency, method, accountNumber, accountName, idempotencyKey]
      properties:
        idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
        merchantId: { type: integer, minimum: 1 }
        amount: { type: number, exclusiveMinimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        method: { type: string }
        accountNumber: { type: string, minLength: 3 }
        accountName: { type: string, minLength: 1 }
        bankCode: { type: string }
        bankName: { type: string }
    PayoutMethodsResponse:
      type: object
      required: [currency, available, minimumWithdrawal, fee, methods]
      properties:
        currency: { type: string }
        available: { type: boolean }
        minimumWithdrawal: { type: number }
        fee:
          type: object
          required: [type]
          properties:
            type: { type: string, enum: [flat, percent] }
            amount: { type: number }
            percent: { type: ["number", "null"] }
            floor: { type: ["number", "null"] }
        methods:
          type: array
          items: { $ref: "#/components/schemas/PayoutMethod" }
    PayoutMethod:
      type: object
      required: [value, label, requiresBankFields]
      properties:
        value: { type: string }
        label: { type: string }
        requiresBankFields: { type: boolean }
    Bank:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
    BankList:
      type: object
      required: [banks]
      properties:
        banks:
          type: array
          items: { $ref: "#/components/schemas/Bank" }
    Settlement:
      type: object
      required: [id, reference, provider, amount, netAmount, currency, status, expectedAt]
      properties:
        id: { type: integer }
        reference: { type: string }
        provider: { type: string, enum: [paystack, payhero, payzaapi] }
        amount: { type: number }
        netAmount: { type: number }
        currency: { type: string }
        status: { type: string, enum: [pending, due, settled, held] }
        expectedAt: { type: string, format: date-time }
        settledAt: { type: ["string", "null"], format: date-time }
        payoutMethod: { type: ["string", "null"] }
    SettlementList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Settlement" }
    Customer:
      type: object
      required: [id, name, email, orderCount, currencySummary]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        orderCount: { type: integer }
        currencySummary: { type: string }
        lastPaymentAt: { type: ["string", "null"], format: date-time }
    CustomerList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
    WebhookEvent:
      type: object
      required: [id, provider, event, status, attempts, receivedAt]
      properties:
        id: { type: integer }
        provider: { type: string }
        event: { type: string }
        reference: { type: ["string", "null"] }
        status: { type: string, enum: [processed, failed, ignored] }
        httpStatus: { type: ["integer", "null"] }
        attempts: { type: integer }
        receivedAt: { type: string, format: date-time }
        lastError: { type: ["string", "null"] }
    WebhookEventList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WebhookEvent" }
...